Refactoring myloader
