+++ src/myloader/myloader_worker_index.h
@@ -20,3 +20,4 @@ void initialize_worker_index(struct configuration *conf);
 void wait_index_worker_to_finish();
 void create_index_shutdown_job(struct configuration *conf);
 void start_optimize_keys_all_tables();
+void enqueue_indexes_if_possible(struct configuration *conf);
